+++ src/rust/iced-x86/src/info/tests/reg_test_parser.rs
@@ -108,8 +108,8 @@ impl Iterator for IntoIter {
 impl IntoIter {
 	fn read_next_test_case(&self, line: String, line_number: u32) -> Result<RegisterInfoTestCase, String> {
 		const_assert_eq!(7, MiscInstrInfoTestConstants::REGISTER_ELEMS_PER_LINE);
-		let elems: Vec<_> = line.splitn(MiscInstrInfoTestConstants::REGISTER_ELEMS_PER_LINE as usize, ',').collect();
-		if elems.len() != MiscInstrInfoTestConstants::REGISTER_ELEMS_PER_LINE as usize {
+		let elems: Vec<_> = line.splitn(MiscInstrInfoTestConstants::REGISTER_ELEMS_PER_LINE, ',').collect();
+		if elems.len() != MiscInstrInfoTestConstants::REGISTER_ELEMS_PER_LINE {
 			return Err(format!("Invalid number of commas: {}", elems.len() - 1));
 		}
