+++ src/rust/iced-x86/src/info/tests/test_parser.rs
@@ -261,8 +261,8 @@ impl Iterator for IntoIter {
 impl IntoIter {
 	fn read_next_test_case(&self, line: String, line_number: u32) -> Result<InstrInfoTestCase, String> {
 		const_assert_eq!(5, MiscInstrInfoTestConstants::INSTR_INFO_ELEMS_PER_LINE);
-		let elems: Vec<_> = line.splitn(MiscInstrInfoTestConstants::INSTR_INFO_ELEMS_PER_LINE as usize, ',').collect();
-		if elems.len() != MiscInstrInfoTestConstants::INSTR_INFO_ELEMS_PER_LINE as usize {
+		let elems: Vec<_> = line.splitn(MiscInstrInfoTestConstants::INSTR_INFO_ELEMS_PER_LINE, ',').collect();
+		if elems.len() != MiscInstrInfoTestConstants::INSTR_INFO_ELEMS_PER_LINE {
 			return Err(format!("Invalid number of commas: {}", elems.len() - 1));
 		}
