+++ src/rust/iced-x86/src/instruction_memory_sizes.rs
@@ -29,7 +29,7 @@ use super::MemorySize;
 // 0 = memory size
 // 1 = broadcast memory size
 #[cfg_attr(feature = "cargo-fmt", rustfmt::skip)]
-pub(crate) static SIZES: [u8; (IcedConstants::NUMBER_OF_CODE_VALUES * 2) as usize] = [
+pub(crate) static SIZES: [u8; (IcedConstants::NUMBER_OF_CODE_VALUES * 2)] = [
 	0,// INVALID
 	MemorySize::UInt8 as u8,// Add_rm8_r8
 	MemorySize::UInt16 as u8,// Add_rm16_r16
